统一为秒计数，修复直播时移问题

+++ android/app/src/com/lecloud/DemoProject/leecoSdk/LeReactPlayerManager.java
@@ -147,7 +147,7 @@ public class LeReactPlayerManager extends SimpleViewManager<LeReactPlayer> {
      * @param seek the msec
      */
     @ReactProp(name = PROP_SEEK)
-    public void setSeek(final LeReactPlayer videoView, final float seek) {
+    public void setSeek(final LeReactPlayer videoView, final int seek) {
         videoView.setSeekTo(seek);
     }
